requirements.txt
alembic==0.9.1
autopep8==1.3
coverage==4.3.4
coveralls==1.1
flake8==3.3.0
Flask==0.12
Flask-HTTPAuth==3.2.2
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
gunicorn==19.7.1
nose==1.3.7
passlib==1.7.1
pep8==1.7.0
psycopg2==2.7.1
python-dotenv==0.6.4
SQLAlchemy==1.1.6
tox==2.7.0
